+++ client/src/Header.js
@@ -1,4 +1,7 @@
-import React from 'react';
+
+import React from 'react'
+
+import axios from 'axios'
 
 import {
   BrowserRouter as Router,
@@ -6,43 +9,52 @@ import {
   Link,
   Redirect,
   withRouter
-} from 'react-router-dom';
+} from "react-router-dom"
 
-import PropTypes from 'prop-types';
+import PropTypes from 'prop-types'
 
-import Button from '@material/react-button';
-import MaterialIcon from '@material/react-material-icon';
+import Button from '@material/react-button'
+import MaterialIcon from '@material/react-material-icon'
 
-import '@material/react-button/dist/button.css';
+import '@material/react-button/dist/button.css'
 import '@material/react-material-icon/dist/material-icon.css';
 
-import './Header.css';
 
-const Codegen = <MaterialIcon icon="cached" />;
+import './Header.css'
 
-const LoginIcon = <MaterialIcon icon="open_in_new" />;
 
-function Header(props) {
-  console.log(props.user);
+const Codegen = <MaterialIcon icon="cached" />
+
+const LoginIcon  = <MaterialIcon icon="open_in_new"/>
+
+function Header (props) {
+  const isAuthenticated = () => !!props.user
+    // try {
+    //   const user = JSON.parse(localStorage.getItem('user'))
+    //   console.log(user)
+    //   return !!user
+    // } catch (error) {
+    //   console.log(error)
+    //   return false
+    // }
+  // }
+
   const Logout = () => (
-    <Route
-      render={({ history }) => (
-        <Button
-          dense
-          color="secondary"
-          outlined
-          onClick={() => {
-            history.push('/logout');
-            localStorage.removeItem('user');
-            window.location.reload();
-          }}
-          className="logout btn"
-        >
-          Logout
-        </Button>
-      )}
-    />
-  );
+    <Route render={({ history }) => (
+      <Button
+        dense
+        outlined
+        onClick={() => {
+          history.push('/logout')
+          localStorage.removeItem('user')
+          window.location.reload()
+        }}
+        className="logout btn"
+      >
+        Logout
+      </Button>  
+    )} />
+  )
 
   // const Logout = () => (
   //   <Button
@@ -52,14 +64,20 @@ function Header(props) {
   //     className="btn"
   //   >
   //     Logout
-  //   </Button>
+  //   </Button> 
   // )
-
+  
   const Login = withRouter(({ history }) => (
-    <Button href="/auth/github" icon={LoginIcon} outlined dense>
+    <Button
+      href="/login/github"
+      className="login"
+      icon={LoginIcon}
+      outlined
+      dense
+    >
       Login with Github
-    </Button>
-  ));
+    </Button>      
+  ))
 
   return (
     <div className="header">
@@ -73,12 +91,12 @@ function Header(props) {
       >
         Generate
       </Button>
-      {!!props.user ? <Logout /> : <Login />}
+      {isAuthenticated() ?  <Logout /> : <Login />}
     </div>
-  );
+  )
 }
 
 Header.propTypes = {
   callback: PropTypes.func
-};
-export default Header;
+}
+export default Header
